server.py

+++ server.py
@@ -1,7 +1,7 @@
 from flask import Flask,redirect,render_template,request,url_for,session
 import mysql.connector
 import base64 
-from datetime import datetime
+from datetime import datetime,timedelta
 
 myconn = mysql.connector.connect(
   host="127.0.0.1",
@@ -25,28 +25,33 @@ reserv_status=None
 # create database hotel;
 # use hotel;
 # create table staff(empid int auto_increment primary key,name varchar(50),email varchar(50),password varchar(50),address varchar(150),designation varchar(15),salary real,gender varchar(1),dob date,phno varchar(15),photo longtext,dateofjoin date);
-# create table reservations(reservation_id int auto_increment primary key,g_count int,rooms_booked int,rooms_id varchar(100),amount real,add_ons varchar(100),addons_price real default 0,gstatus enum('CHECKIN','CHECKOUT','BOOKED'),dates_booked varchar(30));
+# create table reservations(reservation_id int auto_increment primary key,name varchar(25),g_count int,rooms_booked int,rooms_id varchar(100),amount real,add_ons varchar(100),addons_price real default 0,gstatus enum('CHECKIN','CHECKOUT','BOOKED'),dates_booked varchar(30));
 # create table guest(guestid int auto_increment primary key,reservation_id int ,g_name varchar(50),phno varchar(15),country varchar(20),gender varchar(1),room int,image longtext,foreign key(reservation_id) references reservations(reservation_id));
 # create table rooms(roomid int primary key,floorno int,roomstatus ENUM('Vacant', 'Occupied', 'Under Maintenance') DEFAULT 'Vacant',price real,roomtype varchar(10),roomdesc varchar(100),booked_dates longtext,room_image longtext);
-# create table orders(orderid int auto_increment primary key,guestid int,order_list varchar(100),order_update ENUM('Finish','Prep'),o_payment ENUM('Paid','Pending'),price real default 0,foreign key(guestid) references guest(guestid),time datetime);
-# create table amenities(amenity_id int primary key,a_name varchar(50),a_description longtext,head_staff int,a_image longtext,phno varchar(15),floorno int,open_timing time,close_timing time,pricing real,foreign key(head_staff) references staff(empid));
+# create table orders(orderid int auto_increment primary key,reservationid int,order_list varchar(100),order_update ENUM('Finish','Prep'),o_payment ENUM('Paid','Pending'),price real default 0,foreign key(reservationid) references reservations(reservation_id),time datetime);
 
 def refresh():
     db.execute('select roomid,floorno,roomtype,roomdesc,booked_dates,roomstatus,price from rooms;')
     result=db.fetchall()
     global room_status,order_status,staff_status,reserv_status
     room_status=result
-    # db.execute('select orderid,reservationid,order_list,order_update,o_payment,price,time from orders;')
-    db.execute('select * from orders;')
+    db.execute('select * from orders ORDER BY orderid DESC;')
     result=db.fetchall()
     order_status=result
     db.execute('select empid,name,address,designation,salary,phno,dateofjoin from staff;')
     result=db.fetchall()
     print(result)
     staff_status=result
-    db.execute('select * from reservations where gstatus="CHECKIN" OR gstatus="BOOKED";')
+    db.execute('select * from reservations where gstatus="CHECKIN" OR gstatus="BOOKED" ORDER BY reservation_id DESC;')
     result=db.fetchall()
     reserv_status=result
+    
+@app.route('/error')
+def error():
+    global reservationid,checkoutreswid,room_status,order_status,staff_status,reserv_status
+    refresh()
+    return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
+
 
 @app.route('/')
 def home():
@@ -71,8 +76,7 @@ def login():
                 return render_template('options.html')
             else:
                 refresh()
-                return render_template('manager.html',name=session['name'],roomdetails=room_status,ordered=order_status,staffs=staff_status)
-            # return render_template('login.html')
+                return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
         else:
             return redirect(url_for('home'))
     else:
@@ -102,40 +106,68 @@ def addguest():
         image=request.files['file']
         binary_image = image.read() 
         encoded_image = base64.b64encode(binary_image).decode('utf-8')
-        # db.execute('insert into guest (reservation_id,g_name,phno,country,gender,room,image) values (%d,%s,%s,%s,%s,%s,%s);',(reservation,name,phno,country,gender,roomid,encoded_image))
-        # myconn.commit()
         query = 'INSERT INTO guest (reservation_id, g_name, phno, country, gender, room, image) VALUES (%s, %s, %s, %s, %s, %s, %s);'
         values = (reservation, name, phno, country, gender, roomid, encoded_image)
-        # Execute the query
         db.execute(query, values)
-        db.execute('update rooms set roomstatus="Occupied" where roomid=%s',(roomid,))
+        db.execute('update rooms set roomstatus="Occupied" where roomid=%s ',(roomid,))
         db.execute('update reservations set gstatus="CHECKIN" where reservation_id=%s',(reservation,))
         myconn.commit()
         refresh()
-        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status)
-    
+        return render_template('manager.html',name=session['name'],latestreserv=reservation,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
+               
 @app.route('/reservations',methods=['GET','POST'])
 def reservation():
     if request.method=='POST':
+        gname=request.form.get('gname')
         gcount=int(request.form.get('count'))
         totalroom=int(request.form.get('trid'))
         roomid=request.form.get('rid')
-        amount=int(request.form.get('amt'))
-        # datefrom = datetime.strptime(request.form.get('from'), '%Y-%m-%d')
-        # datetill = datetime.strptime(request.form.get('till'), '%Y-%m-%d')
-        # fullstr = f"{datefrom},{datetill}"
+        # amount=int(request.form.get('amt'))
         datefrom=request.form.get('from')
         datetill=request.form.get('till')
-        fullstr=datefrom+","+datetill
-        newstr="BOOKED"
-        print(gcount)
-        print(totalroom)
-        print(roomid)
-        print(amount)
-        print(newstr)
-        print(fullstr)
-        query = 'INSERT INTO reservations (g_count, rooms_booked, rooms_id, amount, gstatus, dates_booked) VALUES (%s, %s, %s, %s, %s, %s)'
-        values = (gcount, totalroom, roomid, amount, newstr, fullstr)
+        # fullstr=datefrom+","+datetill
+        newstr="BOOKED" 
+        roomids = [rid.strip() for rid in roomid.split(',')]
+        datestring=""
+        mainstring=""
+        amount=0
+        for ids in roomids:
+            counter=0
+            db.execute('select * from rooms where roomid=%s and roomstatus="Vacant"',(ids,))
+            a=db.fetchone()
+            if a:
+                db.execute('select booked_dates from rooms where roomid=%s;',(ids,))
+                result=db.fetchone()
+                dates = result[0]
+                datefrom1 = datetime.strptime(datefrom, "%Y-%m-%d")
+                datetill1 = datetime.strptime(datetill, "%Y-%m-%d")
+
+                current_date = datefrom1
+                
+                while current_date < datetill1:
+                    print(current_date.strftime("%Y-%m-%d"))
+                    current_date += timedelta(days=1)
+                    datestring+=current_date.strftime("%Y-%m-%d")+","
+                    counter+=1
+                print(datestring)
+                db.execute('select price from rooms where roomid=%s;',(ids,))
+                result=db.fetchone()
+                amount+=result[0]*counter
+                if dates:
+                    if datefrom in dates or datetill in dates:
+                        myconn.rollback()
+                        return render_template("error.html",message="ROOM NOT AVAILABLE!")     
+                    dates+=" AND "+datestring
+                else:
+                    dates=datestring
+                db.execute('update rooms set booked_dates=%s where roomid=%s;',(dates,ids,))
+                mainstring=datestring
+                datestring=""
+            else:
+                myconn.rollback()
+                return render_template("error.html",message="ROOM NOT VACANT!")
+        query = 'INSERT INTO reservations (name,g_count, rooms_booked, rooms_id, amount, gstatus, dates_booked) VALUES (%s,%s, %s, %s, %s, %s, %s)'
+        values = (gname,gcount, totalroom, roomid, amount, newstr, mainstring)
         db.execute(query, values)
         myconn.commit()
         db.execute('select max(reservation_id) from reservations')
@@ -144,27 +176,9 @@ def reservation():
         global reservationid
         reservationid=result[0]
         print(reservationid)
-        roomids = [rid.strip() for rid in roomid.split(',')]
-        for ids in roomids:
-            db.execute('select booked_dates from rooms where roomid=%s;',(ids,))
-            result=db.fetchone()
-            dates = result[0] if result else ""
-    
-            if dates:
-                dates += " AND " + fullstr
-            else:
-                dates = fullstr
-            # dates=""
-            # print(result)
-            # if result:
-            #     dates=result[0]
-            #     dates = (dates + " AND " + fullstr)
-            # else:
-            #     dates = (dates + fullstr)
-            db.execute('update rooms set booked_dates=%s where roomid=%s;',(dates,ids,))
         myconn.commit()
         refresh()
-        return render_template('manager.html',name=session['name'],latestreserv=result[0],roomdetails=room_status,ordered=order_status,staffs=staff_status)
+        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
         
     
 @app.route('/checkout',methods=['GET','POST'])
@@ -205,18 +219,22 @@ def billed():
             db.execute('select dates_booked from reservations where reservation_id=%s;',(checkoutreswid,))
             result=db.fetchone()
             removebookdate=result[0]
-            removebookdate=" AND "+removebookdate
+            # removebookdate=" AND "+removebookdate
             db.execute('select booked_dates from rooms where roomid=%s;',(roomid,))
             result=db.fetchone()
             resbookdate=result[0]
             if removebookdate in resbookdate:
-                removebookdate = resbookdate.replace(removebookdate[0], "") 
-            db.execute('update rooms set booked_dates=%s where roomid=%s;',(resbookdate,roomid,))
+                removebookdate = resbookdate.replace(removebookdate, "") 
+            if "AND" in removebookdate:
+                removebookdate = removebookdate.replace("AND", "")
+            if removebookdate=="":
+                removebookdate=None        
+            db.execute('update rooms set booked_dates=%s where roomid=%s;',(removebookdate,roomid,))
         myconn.commit()
         db.execute('update orders set o_payment="Paid" where reservationid=%s;',(checkoutreswid,))
         myconn.commit()
         refresh()
-        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status)
+        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
 
 
 @app.route('/ordersubmit',methods=['GET','POST'])
@@ -228,11 +246,15 @@ def ordersubmit():
         print(resid)
         print(orderl)
         print(price)
-        db.execute('insert into orders (reservationid,order_list,order_update,o_payment,price,time) values(%s,%s,"Prep","Pending",%s,NOW())',(resid,orderl,price,))
-        myconn.commit()
-        refresh()
-        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status)
- 
+        db.execute('select reservation_id from reservations where reservation_id=%s',(resid,))
+        result=db.fetchone()
+        if result:
+            db.execute('insert into orders (reservationid,order_list,order_update,o_payment,price,time) values(%s,%s,"Prep","Pending",%s,NOW())',(resid,orderl,price,))
+            myconn.commit()
+            refresh()
+            return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
+        else:
+            return render_template("error.html",message="INVALID RESERVATION ID!")
  
 @app.route('/roomupdate',methods=['GET','POST'])
 def roomupdate():
@@ -250,7 +272,7 @@ def roomupdate():
             db.execute('update rooms set price=%s where roomid=%s;',(price,roomid,))
         myconn.commit()
         refresh()
-        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status)
+        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
 
 
 @app.route('/addstaff',methods=['GET','POST'])
@@ -274,7 +296,7 @@ def addstaff():
         myconn.commit()
         refresh()
         
-        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status)
+        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
 
 
 @app.route('/updatestaff',methods=['GET','POST'])
@@ -282,17 +304,24 @@ def updatestaff():
     if request.method=='POST':
         eid=request.form.get('eid')
         button=request.form.get('buttons')
-        if button=='update':
-            dropdown=request.form.get('selectField')
-            text=request.form.get('addField')
-            db.execute('update staff set %s=%s where empid=%s;',(text,dropdown,eid,))
-            myconn.commit()
+        db.execute('select empid from staff where empid=%s',(eid,))
+        result=db.fetchone()
+        if result:
+            if button=='update':
+                dropdown=request.form.get('selectField')
+                text=request.form.get('addField')
+                s=f"update staff set {dropdown}=%s where empid=%s;"
+                data=(text,eid,)
+                db.execute(s,data)
+                myconn.commit()
+            else:
+                db.execute('delete from staff where empid=%s;',(eid,))
+                myconn.commit()
+            refresh()
+            return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status,reservations=reserv_status)
         else:
-            db.execute('delete from staff where empid=%s;',(eid,))
-            myconn.commit()
-        refresh()
-        return render_template('manager.html',name=session['name'],latestreserv=reservationid,roomdetails=room_status,ordered=order_status,staffs=staff_status)
-
+            return render_template("error.html",message="INVALID EMPLOYEE ID!")
+            
         
  
 @app.route('/chef',methods=['GET','POST'])
@@ -304,7 +333,7 @@ def chef():
             result=db.fetchall()
             return render_template('orders.html',orders=result)
         elif button=='orderhistory':
-            db.execute('select orderid,reservationid,order_list,order_update,o_payment,price,time from orders')
+            db.execute('select orderid,reservationid,order_list,order_update,o_payment,price,time from orders ORDER BY orderid DESC;')
             result=db.fetchall()
             print(result)
             return render_template('orderhistory.html',orders=result)
